+++ flatsurf/geometry/polygon.py
@@ -44,6 +44,7 @@ from sage.structure.sequence import Sequence
 
 
 from .matrix_2x2 import angle
+from .subfield import number_field_elements_from_algebraics
 
 # we implement action of GL(2,K) on polygons
 
@@ -1803,37 +1804,6 @@ class ConvexPolygons(Polygons):
                     raise ValueError("the polygon does not close up")
 
         return self.element_class(self, vertices, check)
-
-def number_field_elements_from_algebraics(elts, name='a'):
-    r"""
-    The native Sage function ``number_field_elements_from_algebraics`` currently
-    returns number field *without* embedding. This function return field with
-    embedding!
-
-    EXAMPLES::
-
-        sage: from flatsurf.geometry.polygon import number_field_elements_from_algebraics
-        sage: z = QQbar.zeta(5)
-        sage: c = z.real()
-        sage: s = z.imag()
-        sage: number_field_elements_from_algebraics((c,s))
-        (Number Field in a with defining polynomial y^4 - 5*y^2 + 5 with a = 1.902113032590308?,
-         [1/2*a^2 - 3/2, 1/2*a])
-    """
-    # case when all elements are rationals
-    if all(x in QQ for x in elts):
-        return QQ, [QQ(x) for x in elts]
-
-    # general case
-    from sage.rings.qqbar import number_field_elements_from_algebraics
-    from sage.rings.number_field.number_field import NumberField
-    field,elts,phi = number_field_elements_from_algebraics(elts, minimal=True)
-
-    polys = [x.polynomial() for x in elts]
-    K = NumberField(field.polynomial(), name, embedding=AA(phi(field.gen())))
-    gen = K.gen()
-
-    return K, [x.polynomial()(gen) for x in elts]
 
 class EquiangularConvexPolygons:
     r"""
